src: render_group: revert "decrease size of bitmap structure"
This reverts commit a9f1cf5ab8fb633f75b07562781a473879409f61.

+++ include/handmadehero/render_group.h
@@ -36,10 +36,10 @@ struct bitmap {
   struct v2 alignPercentage;
   f32 widthOverHeight;
 
-  u16 width;
-  u16 height;
+  u32 width;
+  u32 height;
   // TODO: get rid of stride
-  s16 stride;
+  s32 stride;
   void *memory;
 };
